// src/deviceModule/common/permissiion.type.ts
export type PermissionTypes = {
  READ_CALENDAR: string;
  WRITE_CALENDAR: string;
  CAMERA: string;
  READ_CONTACTS: string;
  WRITE_CONTACTS: string;
  GET_ACCOUNTS: string;
  ACCESS_FINE_LOCATION: string;
  ACCESS_COARSE_LOCATION: string;
  RECORD_AUDIO: string;
  READ_PHONE_STATE: string;
  CALL_PHONE: string;
  ADD_VOICEMAIL: string;
  USE_SIP: string;
  PROCESS_OUTGOING_CALLS: string;
  READ_CALL_LOG: string;
  WRITE_CALL_LOG: string;
  SEND_SMS: string;
  RECEIVE_SMS: string;
  READ_SMS: string;
  RECEIVE_WAP_PUSH: string;
  RECEIVE_MMS: string;
  WRITE_EXTERNAL_STORAGE: string;
  READ_EXTERNAL_STORAGE: string;
  BODY_SENSORS: string;
}